# docs/api/openapi.admin.yaml
openapi: 3.1.0
info:
  title: GreyGray Admin BFF
  version: "1.0.0"
  description: |
    團隊內部 BFF。前面擋著 Cloudflare Access（Zero Trust），
    但**不可以因此省掉應用層的角色檢查**——Access 管的是「誰進得來」，
    不是「進來的人能做什麼」。

    通則見 `docs/05-API契約.md`。與 Storefront 的差別：

    - cookie 名稱 `gg_admin_session`，session 不共用
    - 每個端點標註所需最低角色（`x-required-role`）
    - 金額、enum、錯誤格式與 Storefront 完全相同
servers:
  - url: http://localhost:5001
    description: 開發機
  - url: https://admin.greygray.tw
    description: 正式（經 Cloudflare Access ＋ cloudflared）
...
security:
  - adminSessionCookie: []
paths:
...
  /v1/support/tickets:
    get:
      tags: [support]
      x-required-role: ReadOnly
      operationId: listSupportTickets
      summary: 客服訊息列表
      description: |
        前台右下角小幫手收到的留言（ADR-040）。游標式分頁，**不回總筆數**。
        預設由新到舊。
      parameters:
        - name: status
          in: query
          description: 只看某個狀態；省略為全部。**未結案的優先看**。
          schema: { $ref: '#/components/schemas/SupportTicketStatus' }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicketPage' }
  /v1/support/tickets/{ticketId}:
    get:
      tags: [support]
      x-required-role: ReadOnly
      operationId: getSupportTicket
      summary: 單一客服訊息
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/support/tickets/{ticketId}/resolve:
    post:
      tags: [support]
      x-required-role: Operator
      operationId: resolveSupportTicket
      summary: 標記已處理
      description: |
        ⚠ `Operator` 與 `Accountant` 是**平行角色**，不是等級高低——
        權限判斷不要用 enum 大小比較。
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                staffNote: { type: [string, 'null'], maxLength: 2000 }
      responses:
        '200':
          description: 已標記
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`support.already-resolved`（別人剛處理過）'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
components:
  securitySchemes:
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: gg_admin_session
...
  responses:
...
    NotFound:
      description: 不存在
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
    Id:
      type: string
      pattern: '^[0-9a-f]{32}$'
      description: UUIDv7，32 字元十六進位，無連字號。
...
    ProblemDetails:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: [string, 'null'] }
        instance: { type: [string, 'null'] }
        code: { type: string }
        traceId: { type: [string, 'null'] }
        errors:
          type: [object, 'null']
          additionalProperties:
            type: array
            items: { type: string }
...
    SupportTicketStatus:
      type: string
      enum: [open, resolved]
    SupportTicket:
      type: object
      required: [id, status, message, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        status: { $ref: '#/components/schemas/SupportTicketStatus' }
        message: { type: string, maxLength: 2000 }
        contactEmail: { type: [string, 'null'], maxLength: 200 }
        contactPhone: { type: [string, 'null'], maxLength: 20 }
        menuPath:
          description: |
            客人在小幫手裡走過的選單路徑。**客服要先看這個**——
            只有一句「不能用」而不知道他卡在哪一題，回覆只能猜。
          type: array
          maxItems: 5
          items: { type: string, maxLength: 100 }
        orderId:
          oneOf:
            - { $ref: '#/components/schemas/Id' }
            - { type: 'null' }
        customerId:
          description: 匿名留言為 null。
          oneOf:
            - { $ref: '#/components/schemas/Id' }
            - { type: 'null' }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        resolvedBy: { type: [string, 'null'] }
        staffNote: { type: [string, 'null'], maxLength: 2000 }
    SupportTicketPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SupportTicket' }
        nextCursor:
          description: null 代表沒有下一頁。**不回總筆數**（游標式分頁的慣例）。
          type: [string, 'null']